Update, fix for windows

+++ _targets.R
@@ -315,34 +315,41 @@ list(
   
   # tarchetypes::tar_render(report, path = "README.Rmd", params = list(zones, rnet)),
   tar_target(upload_data, {
-  
+
     length(combined_network)
     length(r_commute)
     commit = gert::git_log(max = 1)
-    v = paste0("v", save_outputs, "_commit_", commit$commit)
-    v = gsub(pattern = " |:", replacement = "-", x = v)
-    setwd("outputdata")
-    f = list.files(path = ".", pattern = "Rds")
-    # Piggyback fails with error message so commented and using cust
-    # piggyback::pb_upload(f)
-    msg = glue::glue("gh release create {v} --generate-notes")
-    message("Creating new release and folder to save the files: ", v)
-    dir.create(v)
-    message("Going to try to upload the following files: ", paste0(f, collapse = ", "))
-    message("With sizes: ", paste0(fs::file_size(f), collapse = ", "))
-    system(msg)
-    for(i in f) {
-      gh_release_upload(file = i, tag = v)
-      # Move into a new directory
-      file.copy(from = v, to = file.path(v, i))
+    message("Commit: ", commit)
+    
+    if(Sys.info()[['sysname']] == "Linux") {
+      v = paste0("v", save_outputs, "_commit_", commit$commit)
+      v = gsub(pattern = " |:", replacement = "-", x = v)
+      setwd("outputdata")
+      f = list.files(path = ".", pattern = "Rds")
+      # Piggyback fails with error message so commented and using cust
+      # piggyback::pb_upload(f)
+      msg = glue::glue("gh release create {v} --generate-notes")
+      message("Creating new release and folder to save the files: ", v)
+      dir.create(v)
+      message("Going to try to upload the following files: ", paste0(f, collapse = ", "))
+      message("With sizes: ", paste0(fs::file_size(f), collapse = ", "))
+      system(msg)
+      for(i in f) {
+        gh_release_upload(file = i, tag = v)
+        # Move into a new directory
+        file.copy(from = v, to = file.path(v, i))
+      }
+      message("Files stored in output folder: ", v)
+      message("Which contains: ", paste0(list.files(v), collapse = ", "))
+      # For rds based version:
+      # For specific version:
+      # system("gh release create v0.0.1 --generate-notes")
+      file.remove(f)
+      setwd("..")
+    }  else {
+      message("gh command line tool not available")
+      message("Now create a release with this version number and upload the files")
     }
-    message("Files stored in output folder: ", v)
-    message("Which contains: ", paste0(list.files(v), collapse = ", "))
-    # For rds based version:
-    # For specific version:
-    # system("gh release create v0.0.1 --generate-notes")
-    file.remove(f)
-    setwd("..")
   })
   # tar_source(files = "data-raw/test-tiles.R") # how to source script as target?
 )
